+++ lib/src/widget/clip/clip_rect.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/widgets.dart';
+
+extension ClipRectWidgetExtensions on Widget {
+  Widget clipRRect({
+    BorderRadiusGeometry borderRadius = BorderRadius.zero,
+    CustomClipper<RRect>? clipper,
+    Clip clipBehavior = Clip.antiAlias,
+  }) {
+    return ClipRRect(
+      borderRadius: borderRadius,
+      clipper: clipper,
+      clipBehavior: clipBehavior,
+      child: this,
+    );
+  }
+
+  Widget clipOval({
+    CustomClipper<Rect>? clipper,
+    Clip clipBehavior = Clip.antiAlias,
+  }) {
+    return ClipOval(
+      clipper: clipper,
+      clipBehavior: clipBehavior,
+      child: this,
+    );
+  }
+
+  Widget clipCircle({
+    CustomClipper<Rect>? clipper,
+    Clip clipBehavior = Clip.antiAlias,
+  }) =>
+      clipOval(
+        clipper: clipper,
+        clipBehavior: clipBehavior,
+      );
+}